[MODMO-2]. Apply review recommendations
